requirements.txt
PyGObject~=3.48.2
psycopg2-binary~=2.9.9
argparse~=1.4.0
setuptools~=68.2.0